refactor: combine service in single file
