--- internal/services/dispatch/sendDispatch.go
@@ -1,7 +1,0 @@
-package ds
-
-import "context"
-
-func (s *dispatchService) SendDispatch(ctx context.Context, dispatch string) error {
-	return nil
-}
